Add files via upload
The Iris flower data set or Fisher's Iris data set is a multivariate data set introduced by the British statistician and biologist Ronald Fisher.The Iris dataset is a classic dataset for classification, machine learning, and data visualization. In this notebook I'm going to do an end to end project of Machine Learning in which we will take SepalLength, SepalWidth, PetalLength, PetalWidth an input and will predict the species of the flower.

 After completing this project I have learned how to do a project in data science from collecting data to understanding the working of these particular algorithms.

I am still learning! :)